Refactor exception documentation for cache methods
Removes the requirement to throw \Psr\SimpleCache\InvalidArgumentException when a parameter "is neither an array nor a Traversable". With the `iterable` definition on the parameter, it is functionally *impossible* to do so. 

PHP will natively throw a TypeError before the code is given any chance to throw its own exception